Implement BTBX with aliasing tracking
In this commit we have a clean implementation of BTBX including alias
tracking in the CPU side of the branch predictor. We are now also
dumping out all raw data that we can normalise during aggregation
instead of having percentage data only.

Still unclear as to why we observe worse performance on certain region
based btbs than the baselines with less bits for the tag.